Move the location of the rating-reviews
