+++ packages/review/src/components/shorten-list/index.ts
@@ -1,2 +1,3 @@
 export * from './popular-reviews'
 export * from './latest-reviews'
+export * from './rating-reviews'
